Implemented writing cached tracklists to .jspf (at this point unused)
JSPF will be used for caching tracklists, as we want to remove the
dependency on libxml-2.0.
A JSON format was chonsen, simply because JSON is required nevertheless
to be able to read data from soundcloud.com.
The XML parser is, at this point, only used by xspf.c and therefore
the perfect candidate to be replaced.